VCC-2970 VCC-2968 Add a select-only variant; add additional keyboard interaction
The input is replaced with a `div`, then all of the appropriate interactions and attributes are added to the `div`.

Notably, this also brings a lot of `select`-specific behavior to the `div`. Even though the text input is disabled and not visible, we still have to use familiar patterns.