Set `aria-labelledby` on transformed or `selectOnly` comboboxes
When scanning with Axe, there was an error that the `selectOnly` combobox was missing a label. The `<label>` had its `for` attribute set, but because of the element switching/transformation, it was not being recognized and paired. To fix this, I added an `aria-labelledby` attribute to the combobox. The `id` of the `<label>` will be used if it exists; otherwise, a random `id` will be generated.