Use new helper methods, minor refactor
Minor refactor to organize config information in the constructor. Check for internal calls to make sure that Combobo is ready and has initialized. This also adds better (and more consistent) support for `className` to ensure that APIs are compatible with allowing custom class names on options/optgroups. Optgroups can also be added to the top or bottom, rather than always on the bottom.