fix: Filter comboboxes by their visible text instead of their value (PE-374)
When a combobox is not select-only (e.g. it allows text input), and the visible text had a mismatch with the `data-value`, the visible text could not be searched. This gets the visible text from `innerText` or `textContent` and searches for that instead of the value.

Two new tests added!